refactor: clean obsolete paint and window config paths
