+++ src/apps/paint/types.ts
@@ -1,0 +1,4 @@
+export interface Filter {
+  name: string;
+  apply: (canvas: HTMLCanvasElement) => void;
+}
